Convert all injectables for AppPaths to be InitializableState
- Simplifies the number of injectable and consolidates the state
  checking
