+++ src/main/app-paths/app-paths-request-channel-listener.injectable.ts
@@ -3,15 +3,15 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { appPathsChannel } from "../../common/app-paths/app-paths-channel";
-import appPathsInjectable from "../../common/app-paths/app-paths.injectable";
 import { getRequestChannelListenerInjectable } from "../utils/channel/channel-listeners/listener-tokens";
+import appPathsInjectable from "./impl.injectable";
 
 const appPathsRequestChannelListenerInjectable = getRequestChannelListenerInjectable({
   channel: appPathsChannel,
   handler: (di) => {
     const appPaths = di.inject(appPathsInjectable);
 
-    return () => appPaths;
+    return () => appPaths.get();
   },
 });
